task11

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentService;
 
 /**
  * 削除コントローラー
@@ -23,6 +24,9 @@ public class DeleteBookController {
     
     @Autowired
     private BooksService booksService;
+    
+    @Autowired
+    private RentService rentService;
 
     /**
      * 対象書籍を削除する
@@ -36,15 +40,26 @@ public class DeleteBookController {
     @RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
     public String deleteBook(
             Locale locale,
-            @RequestParam("bookId") Integer bookId,
+            @RequestParam("bookId") int bookId,
             Model model) {
         logger.info("Welcome delete! The client locale is {}.", locale);
 
-        booksService.deleteBook(bookId);
-        model.addAttribute("bookList", booksService.getBookList());
         
+        int deleteId2 = rentService.delete(bookId);
+        
+        
+        if(deleteId2 == 0) {
+        	booksService.deleteBook(bookId);
+        }else {
+        	model.addAttribute("deleteError","貸出し中です。");
+        	model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+        	return"details";
+        }
+        
+        
+        model.addAttribute("bookList", booksService.getBookList());
         return "home";
-
+        
     }
 
 }
